Move ready to publish to modal, add delete module modal
